# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "DiagnosticsResponse": {
        "additionalProperties": false,
        "properties": {
          "interrupted_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterruptedRunResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "limits": {
            "$ref": "#/components/schemas/LimitsResponse"
          },
          "open_jobs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/OpenJobResponse"
            },
            "title": "Open Jobs",
            "type": "array"
          },
          "parse_error_count": {
            "title": "Parse Error Count",
            "type": "integer"
          },
          "repository_id": {
            "title": "Repository Id",
            "type": "string"
          },
          "skipped_by_reason": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Skipped By Reason",
            "type": "object"
          },
          "snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Id"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "repository_id",
          "snapshot_id",
          "skipped_by_reason",
          "parse_error_count",
          "limits",
          "warnings"
        ],
        "title": "DiagnosticsResponse",
        "type": "object"
      },
...
      "FilesResponse": {
        "additionalProperties": false,
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileEntry"
            },
            "title": "Files",
            "type": "array"
          },
          "repository_id": {
            "title": "Repository Id",
            "type": "string"
          },
          "snapshot": {
            "$ref": "#/components/schemas/SnapshotReference"
          }
        },
        "required": [
          "repository_id",
          "snapshot",
          "files"
        ],
        "title": "FilesResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegisterRepositoryBody": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "path": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "RegisterRepositoryBody",
        "type": "object"
      },
...
      "RepositoryResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "repository_id": {
            "title": "Repository Id",
            "type": "string"
          }
        },
        "required": [
          "repository_id",
          "display_name",
          "created_at"
        ],
        "title": "RepositoryResponse",
        "type": "object"
      },
...
      "StatusResponse": {
        "additionalProperties": false,
        "properties": {
          "file_count": {
            "title": "File Count",
            "type": "integer"
          },
          "parse_error_count": {
            "title": "Parse Error Count",
            "type": "integer"
          },
          "repository_id": {
            "title": "Repository Id",
            "type": "string"
          },
          "snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SnapshotReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "symbol_count": {
            "title": "Symbol Count",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "repository_id",
          "snapshot",
          "file_count",
          "symbol_count",
          "parse_error_count",
          "warnings"
        ],
        "title": "StatusResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CodeAtlas local API",
    "version": "1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/repositories": {
      "get": {
        "operationId": "list_repositories_v1_repositories_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RepositoryResponse"
                  },
                  "title": "Response List Repositories V1 Repositories Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Repositories",
        "tags": [
          "repositories"
        ]
      },
      "post": {
        "operationId": "register_repository_v1_repositories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRepositoryBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Repository",
        "tags": [
          "repositories"
        ]
      }
    },
    "/v1/repositories/{repository_id}": {
      "delete": {
        "description": "Remove a repository from CodeAtlas. Source files are never touched.\n\nRefuses with `REPOSITORY_HAS_CONVERSATIONS` while conversations exist,\nunless `cascade` is set. The schema would take them silently; a user\nfreeing index space should not lose chat history without saying so.",
        "operationId": "delete_repository_v1_repositories__repository_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          },
          {
            "description": "Also delete the repository's conversations.",
            "in": "query",
            "name": "cascade",
            "required": false,
            "schema": {
              "default": false,
              "description": "Also delete the repository's conversations.",
              "title": "Cascade",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Repository",
        "tags": [
          "repositories"
        ]
      },
      "get": {
        "operationId": "get_repository_v1_repositories__repository_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Repository",
        "tags": [
          "repositories"
        ]
      }
    },
    "/v1/repositories/{repository_id}/diagnostics": {
      "get": {
        "operationId": "repository_diagnostics_v1_repositories__repository_id__diagnostics_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Repository Diagnostics",
        "tags": [
          "repositories"
        ]
      }
    },
    "/v1/repositories/{repository_id}/files": {
      "get": {
        "description": "List the files the active snapshot contains.\n\nPaths are repository-relative, as everywhere else: the absolute root is not\nechoed back, so a screenshot or diagnostic bundle cannot leak it.",
        "operationId": "repository_files_v1_repositories__repository_id__files_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Repository Files",
        "tags": [
          "repositories"
        ]
      }
    },
    "/v1/repositories/{repository_id}/index": {
      "post": {
        "operationId": "index_repository_v1_repositories__repository_id__index_post",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Index Repository",
        "tags": [
          "repositories"
        ]
      }
    },
    "/v1/repositories/{repository_id}/rollback": {
      "post": {
        "description": "Restore the previous snapshot as active.\n\nThe escape hatch for an activation that turned out to be wrong. With no\nprevious snapshot to return to this is a conflict, not a server error.",
        "operationId": "rollback_v1_repositories__repository_id__rollback_post",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotReference"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rollback",
        "tags": [
          "repositories"
        ]
      }
    },
    "/v1/repositories/{repository_id}/semantic-status": {
      "get": {
        "description": "How much of the active snapshot the semantic index covers.\n\nNever refuses for an unindexed or opted-out repository: both are ordinary\nstates with ordinary answers, and a 409 would make a client treat the\ndefault configuration as a fault.",
        "operationId": "semantic_status_v1_repositories__repository_id__semantic_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemanticStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Semantic Status",
        "tags": [
          "repositories"
        ]
      }
    },
    "/v1/repositories/{repository_id}/snapshots/active": {
      "get": {
        "operationId": "active_snapshot_v1_repositories__repository_id__snapshots_active_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotReference"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Active Snapshot",
        "tags": [
          "repositories"
        ]
      }
    },
    "/v1/repositories/{repository_id}/status": {
      "get": {
        "operationId": "repository_status_v1_repositories__repository_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Repository Status",
        "tags": [
          "repositories"
        ]
      }
    },
    "/v1/repositories/{repository_id}/watch": {
      "get": {
        "description": "Report the watch switch and what the watcher is actually doing.",
        "operationId": "get_watch_v1_repositories__repository_id__watch_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Watch",
        "tags": [
          "repositories"
        ]
      },
      "put": {
        "description": "Turn continuous freshness on or off for one repository.\n\nThe decision is persisted, so it survives a restart: turning the watcher off\nis a statement about the repository, not about this process.",
        "operationId": "set_watch_v1_repositories__repository_id__watch_put",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Watch",
        "tags": [
          "repositories"
        ]
      }
    },
...
  }
}